requirements.txt
fastapi==0.124.0
uvicorn==0.30.3
joblib==1.5.2
matplotlib==3.10.7
numpy==2.3.5
pandas==2.3.3
Pillow==12.0.0
pydantic==2.12.5
python-multipart==0.0.9
scikit-learn==1.7.2
scipy==1.16.3
tensorflow-cpu==2.20.0
Flask==3.1.0
requests==2.32.3
